fix(fp-frame): pct_change accepts negative (forward) periods like pandas (br-frankenpandas-p7c3l)
pct_change took periods: usize on Series, DataFrame and both groupby
variants, so pandas' negative pct_change(-1) (forward percent change) was
impossible — inconsistent with shift/diff/pct_change_axis1 (already i64),
and the conformance dispatch explicitly rejected negative DataFrame
periods. Switch the signature to i64 and compute the comparison index as a
signed offset (prev = i - periods; NaN when outside [0, len)), so positive
periods look back and negative look forward, exactly like
current/shift(periods)-1. Honored across the typed-f64 fast path, the
general path, and the per-group transforms; the default pad-fill runs
before the signed shift unchanged.

Conformance dispatch now passes i64 and drops the non-negative guard;
fp-expr call sites cast (the expr AST field stays usize). Verified vs a
6-case live pandas 2.2.3 oracle (forward/back periods + pad-fill/None). 23
pct_change tests green, clippy clean (fp-frame/fp-conformance/fp-expr).

+++ crates/fp-expr/src/lib.rs
@@ -708,7 +708,7 @@ pub fn evaluate(
         }
         Expr::PctChange { expr, periods } => {
             let input = evaluate(expr, context, policy, ledger)?;
-            input.pct_change(*periods).map_err(ExprError::from)
+            input.pct_change(*periods as i64).map_err(ExprError::from)
         }
         Expr::Compare { left, right, op } => {
             evaluate_comparison(left, right, *op, context, policy, ledger)
@@ -1499,7 +1499,7 @@ fn evaluate_delta(
         }
         Expr::PctChange { expr, periods } => {
             let input = evaluate_delta(expr, delta_ctx, delta, policy, ledger)?;
-            input.pct_change(*periods).map_err(ExprError::from)
+            input.pct_change(*periods as i64).map_err(ExprError::from)
         }
         Expr::Compare { left, right, op } => {
             evaluate_delta_comparison(left, right, *op, delta_ctx, delta, policy, ledger)
